FIX: Diet 생성 시 loading Dialog 안사라지는 버그 해결
1. DietId의 타입을 Long으로 변경
    Insert의 결과를 받고 Entity를 반환하기 위해
    Entity의 id 타입을 Long으로 변경함.

2. Result를 사용해 성공 / 실패 전달
    viewModel의 dietResult로 결과를 전달

3. LiveData 관찰로 diet 생성 확인 후 dialog 닫기

+++ app/src/main/java/com/dearmyhealth/data/db/entities/Diet.kt
@@ -24,7 +24,7 @@ import androidx.room.PrimaryKey
 )
 data class Diet(
     @PrimaryKey(autoGenerate = true)
-    val dietId: Int,
+    val dietId: Long,
     val foodCode: String?,
     val user: Int,
     val time: Long,
